Some small fixes
Certificate bundle added to provide by certifi package for the requests library to use for verification.
Usage of the default user-agent of the requests library, removed the disabling of InsecureRequestWarning and timeout set to 10 sec.
It should work without giving the OSError: Could not find a suitable TLS CA certificate bundle